Test insterrupt on UART
Result:
Work correctly except for the "SAID_HELLO" part.
Hello_world message is not received on minicom even when already
connected before restart.
Not sure on how works the atomic memory "ordering" part...

Define interupt function using the coresponding attribute.
As interrupt are hardware related, the attribute need to come from the
device crate. Not from cortex_m_rt crate!

Option are use in Static variable for driver access. Don't if it's the
best practice but it's what is used in rp-board repository exemple.